Add a selector class.
Add a selector class which optimizes some dropdown behavior.  In
particular, it defers the generation of the dropdown's contents until
the first time the user opens the dropdown.  When put to use, this
helps page load time a lot on pages which have a lot of dropdowns,
such as complicated Bobs recipes with many module options for
factories.