Change how icon tooltips work.
Refactor icons so that they place their toolip in a span container
along side of the icon at the same time instead of placing the tooltip
into the body of the page.  This makes it easier to clean up the
tooltip and not accidenally leak them as pages get updated with new
content.  Now that span contains both, by default, purging the span
purges both the icon and the tooltip.

Also allow users the ability to specify which node the tooltip is
placed in.  When creating tooltips for dropdowns, we want to create
the tooltip in the dropdown's wrapper, not the dropdown contents
themselves.  Placing the tooltip under the dropdown itself will cause
it to be clipped by the dropdown contents.