Allow factory selection in the factory tab
Allow users to select which factory they specifically want to use in a solution
from the factory tab.

TODO: debug why the solution does not seem to be updating the number of
      factories properly.
TODO: fix the concept of "minimum" factory.  With bob's mods, this might be
      better done as "default for factory category"
TODO: persist the factory selection in the URL