Track mining recipes separately
Prepend "mining-" to mining recipes created from resources in order to
keep them separate from other recipes.  Bob's mods (when playing with
the ore mod enabled) currently lists "sulfur" as a resource which can
be mined, in addition to having up to 3 possible recipes for creating
sulfur, one of which is named simply "sulfur" which conflicts with the
name which was being auto-generated.

Also, start to track the localized name provided with either the
resource, or with the recipe itself.  We will use this in a future
change (when available) for name generation.